move NPUCachingAllocatorHelper
