+++ npu/core/sys_ctrl/npu_sys_ctrl.cpp
@@ -1,9 +1,9 @@
 #include "npu/core/sys_ctrl/npu_sys_ctrl.h"
 #include "csrc/npu/NPUCachingAllocator.h"
+#include "csrc/npu/NPUCachingAllocatorHelper.h"
 #include "csrc/npu/NPUFunctions.h"
 #include "csrc/npu/NPUStream.h"
 #include "npu/acl/include/acl/acl_op_compiler.h"
-#include "npu/core/NPUCachingAllocatorHelper.h"
 #include "npu/core/NpuVariables.h"
 #include "npu/core/npu_log.h"
 #include "npu/core/register/OptionRegister.h"
